Chore: moves safe* functions to grafana/data
