+++ public/app/features/variables/inspect/utils.ts
@@ -1,10 +1,10 @@
+import { DataLinkBuiltInVars, safeStringifyValue } from '@grafana/data';
+
 import { variableAdapters } from '../adapters';
 import { DashboardModel, PanelModel } from '../../dashboard/state';
 import { isAdHoc } from '../guard';
-import { safeStringifyValue } from '../../../core/utils/explore';
 import { VariableModel } from '../types';
 import { containsVariable, variableRegex, variableRegexExec } from '../utils';
-import { DataLinkBuiltInVars } from '@grafana/data';
 
 export interface GraphNode {
   id: string;
