+++ public/app/features/variables/utils.ts
@@ -1,12 +1,11 @@
 import { isArray, isEqual } from 'lodash';
-import { ScopedVars, UrlQueryMap, UrlQueryValue, VariableType } from '@grafana/data';
+import { safeStringifyValue, ScopedVars, UrlQueryMap, UrlQueryValue, VariableType } from '@grafana/data';
 import { getTemplateSrv } from '@grafana/runtime';
 
 import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from './constants';
 import { QueryVariableModel, VariableModel, VariableRefresh } from './types';
 import { getTimeSrv } from '../dashboard/services/TimeSrv';
 import { variableAdapters } from './adapters';
-import { safeStringifyValue } from 'app/core/utils/explore';
 import { StoreState } from '../../types';
 import { getState } from '../../store/store';
 import { TransactionStatus } from './state/transactionReducer';
