+++ public/app/plugins/datasource/prometheus/datasource.ts
@@ -20,6 +20,7 @@ import {
   TimeRange,
   DataFrame,
   dateTime,
+  safeStringifyValue,
 } from '@grafana/data';
 import {
   BackendSrvRequest,
@@ -31,7 +32,6 @@ import {
   toDataQueryResponse,
 } from '@grafana/runtime';
 
-import { safeStringifyValue } from 'app/core/utils/explore';
 import { getTimeSrv, TimeSrv } from 'app/features/dashboard/services/TimeSrv';
 import { getTemplateSrv, TemplateSrv } from 'app/features/templating/template_srv';
 import addLabelToQuery from './add_label_to_query';
